requirements.txt
sktime
pmdarima
pandas
matplotlib
seaborn
plotly
scikit-learn
yfinance
lightgbm
prophet
statsmodels
streamlit
geopandas
folium
xmltodict
streamlit-option-menu
millify
pingouin
jupyterlab
PyMuPDF
openpyxl==3.1.5
requests
Pillow
streamlit-autorefresh